task 3 - fix checked clauses counting

+++ SatDotNet.Core/Interfaces.cs
@@ -31,7 +31,7 @@ namespace SatDotNet.Core
         bool IsSafisfied { get; }
         bool IsUnsatisfiable { get; }
         int UnitPropagationSteps { get; }
-        int CheckedClauses { get; }
+        long CheckedClauses { get; }
         IEnumerable<ILiteral> GetAssignment();
     }
